# docs/swagger.yaml
basePath: /
definitions:
  model.CreateLogRequest:
    properties:
      level:
        type: string
      message:
        type: string
      service:
        type: string
    type: object
  model.Log:
    properties:
      created_at:
        type: string
      id:
        type: string
      level:
        type: string
      message:
        type: string
      service:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Logging service with REST + gRPC
  title: Log Service API
  version: "1.0"
paths:
  /logs:
    post:
      consumes:
      - application/json
      description: Insert a log entry
      parameters:
      - description: Log input
        in: body
        name: log
        required: true
        schema:
          $ref: '#/definitions/model.CreateLogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Log'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new log
      tags:
      - logs
  /logs/batch:
    post:
      consumes:
      - application/json
      description: Insert multiple log entries (max 500)
      parameters:
      - description: List of logs
        in: body
        name: logs
        required: true
        schema:
          items:
            $ref: '#/definitions/model.CreateLogRequest'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            items:
              $ref: '#/definitions/model.Log'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create logs in batch
      tags:
      - logs
  /logs/search:
    get:
      description: Filter logs by level/service/regex
      parameters:
      - description: Log level
        in: query
        name: level
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      - description: Regex filter
        in: query
        name: regex
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Log'
            type: array
      summary: Search logs
      tags:
      - logs
swagger: "2.0"